Options menu

+++ Game Inventory/DigitalGame.h
@@ -7,20 +7,20 @@
 
 class DigitalGame : public Game {
 private:
-    double downloadSize; // Size of the game in GB
+    float downloadSize; // Size of the game in GB
 
 public:
     // Constructors
     DigitalGame();
     DigitalGame(const std::string& title, const std::string& genre, int year, const Platform& platform,
-                double downloadSize);
+                float downloadSize);
     ~DigitalGame() override;
 
     // Setter for downloadSize
-    void setDownloadSize(double size);
+    void setDownloadSize(float size);
 
     // Getter for downloadSize
-    double getDownloadSize() const;
+    float getDownloadSize() const;
 
     // Overridden printInfo method to include downloadSize
     void printInfo() const override;
